# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Top Tin Roofing CRM & Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: stage
          in: query
          schema: { type: string }
        - name: assignedTo
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Lead" }
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
    patch:
      operationId: updateLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
    delete:
      operationId: deleteLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /leads/{id}/stage:
    patch:
      operationId: updateLeadStage
      tags: [leads]
      summary: Move lead to a different pipeline stage
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadStageUpdate" }
      responses:
        "200":
          description: Stage updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
  /leads/{id}/notes:
    post:
      operationId: addLeadNote
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoteInput" }
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }

  # ── CUSTOMERS ──────────────────────────────────────────────────────────────
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: customerId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── CREW MEMBERS ───────────────────────────────────────────────────────────
...
  /materials:
    get:
      operationId: getMaterials
      tags: [materials]
      responses:
        "200":
          description: List of roofing materials
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Material" }
    post:
      operationId: createMaterial
      tags: [materials]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Material" }
  /materials/{id}:
    patch:
      operationId: updateMaterial
      tags: [materials]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Material" }
    delete:
      operationId: deleteMaterial
      tags: [materials]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── ACTIVITY LOGS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, name, email, phone, stage, priority, source, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: ["string", "null"] }
        roofType: { type: ["string", "null"] }
        squareFootage: { type: ["number", "null"] }
        stage: { type: string }
        priority: { type: string }
        source: { type: string }
        assignedTo: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        estimatedValue: { type: ["number", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LeadInput:
      type: object
      required: [name, email, phone, stage, priority, source]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        roofType: { type: string }
        squareFootage: { type: number }
        stage: { type: string }
        priority: { type: string }
        source: { type: string }
        assignedTo: { type: string }
        notes: { type: string }
        estimatedValue: { type: number }
    LeadUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        roofType: { type: string }
        squareFootage: { type: number }
        stage: { type: string }
        priority: { type: string }
        source: { type: string }
        assignedTo: { type: string }
        notes: { type: string }
        estimatedValue: { type: number }
    LeadStageUpdate:
      type: object
      required: [stage]
      properties:
        stage: { type: string }
    NoteInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
        author: { type: string }
    Note:
      type: object
      required: [id, content, createdAt]
      properties:
        id: { type: integer }
        content: { type: string }
        author: { type: ["string", "null"] }
        leadId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }

    # ── CUSTOMER ────────────────────────────────────────────────────────────
...
    Invoice:
      type: object
      required: [id, customerId, status, totalAmount, dueDate, createdAt, updatedAt]
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        status: { type: string }
        subtotal: { type: ["number", "null"] }
        taxAmount: { type: ["number", "null"] }
        discountAmount: { type: ["number", "null"] }
        totalAmount: { type: number }
        paidAmount: { type: ["number", "null"] }
        dueDate: { type: string, format: date }
        paidAt: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
        lineItems: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    InvoiceInput:
      type: object
      required: [customerId, status, totalAmount, dueDate]
      properties:
        customerId: { type: integer }
        projectId: { type: integer }
        status: { type: string }
        subtotal: { type: number }
        taxAmount: { type: number }
        discountAmount: { type: number }
        totalAmount: { type: number }
        paidAmount: { type: number }
        dueDate: { type: string, format: date }
        notes: { type: string }
        lineItems: { type: string }
    InvoiceUpdate:
      type: object
      properties:
        status: { type: string }
        subtotal: { type: number }
        taxAmount: { type: number }
        discountAmount: { type: number }
        totalAmount: { type: number }
        paidAmount: { type: number }
        dueDate: { type: string, format: date }
        paidAt: { type: string, format: date }
        notes: { type: string }
        lineItems: { type: string }

    # ── CREW MEMBER ─────────────────────────────────────────────────────────
...
    Material:
      type: object
      required: [id, name, category, unit, pricePerUnit, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        description: { type: ["string", "null"] }
        unit: { type: string }
        pricePerUnit: { type: number }
        stockQuantity: { type: ["integer", "null"] }
        supplier: { type: ["string", "null"] }
        sku: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MaterialInput:
      type: object
      required: [name, category, unit, pricePerUnit]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        description: { type: string }
        unit: { type: string }
        pricePerUnit: { type: number }
        stockQuantity: { type: integer }
        supplier: { type: string }
        sku: { type: string }
    MaterialUpdate:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        description: { type: string }
        unit: { type: string }
        pricePerUnit: { type: number }
        stockQuantity: { type: integer }
        supplier: { type: string }
        sku: { type: string }

    # ── ACTIVITY LOG ────────────────────────────────────────────────────────
...