# lib/api-zod/src/generated/types/leadStageUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Top Tin Roofing CRM & Operations Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadStageUpdate {
  stage: string;
}