# lib/api-zod/src/generated/types/materialInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Top Tin Roofing CRM & Operations Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialInput {
  /** @minLength 1 */
  name: string;
  category: string;
  description?: string;
  unit: string;
  pricePerUnit: number;
  stockQuantity?: number;
  supplier?: string;
  sku?: string;
}